+++ retire/__init__.py
@@ -1,5 +1,3 @@
 # retire/__init__.py
-from .data import load_dataset
-from .data import load_clean_dataset
-from .data import load_projection
-from .data import load_graph
+
+from retire.explore import Explore
