+++ retire/examples/__init__.py
@@ -1,0 +1,2 @@
+from retire.examples.heatmap_config import heatmap_config
+from retire.examples.dotplot_config import dotplot_config
