--- retire/figures/__init__.py
@@ -1,1 +1,0 @@
-from .plotkit import PlotKit
