+++ retire/resources/results/plant_archetype_matches.csv
@@ -1,0 +1,4 @@
+Plant Name,Group Match,Match to Archetype?,Match Explanation
+Dolet Hills,"{4: 0.94, 5: 0.06}",Perfect Match,Plant's high operating costs and utility cost reduction efforts directly match Group 4's primary driver of high operating costs.
+Martin Drake // South Plant,{7: 1.0},Close Match,"Plant is retiring due to aging and transition to cleaner energy, but were previously in violation of federal air quality regulations and, as such, match the `Air Quality Offenders' Group 7 archetype."
+E D Edwards,{6: 1.0},Partial Match,Plant closed due to legal settlement from environmental violations. This partially aligns with Group 6's `Political Opposition' (manifested as regulatory and legal action) and indirectly with `State-driven climate goals' which often underpin such environmental standards.
